Add updatezoomtile command
This adds a command, 'updatezoomtiles' with the same usage as fullrender, that causes all zoomtiles to be recreated.

This approach is nonideal;

Firstly, it doesn't actually run the zoom-tile generation like fullrender does tile generation. Zoom tile generation is still queued for execution in the background. This is probably the easiest thing to address.

Secondly, it performs a lot of the same processing as normal tile generation except for the final step of rendering and saving. I don't know enough about the codebase yet to know how much this can be improved, but I do know that magically generating a list of zoom tiles to compute is nontrivial. It either needs to iterate over tiles in storage, or chunks in the same way fullrender does to find zoom tiles that need generating. It's possible almost no improvement can be made here.

This needs significant improvement before I'd consider it for merging..